// packages/prisma-oppulence/generated/zod/inputTypeSchemas/NestedEnumFeedbackTypeWithAggregatesFilterSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { FeedbackTypeSchema } from './FeedbackTypeSchema';
import { NestedIntFilterSchema } from './NestedIntFilterSchema';
import { NestedEnumFeedbackTypeFilterSchema } from './NestedEnumFeedbackTypeFilterSchema';

export const NestedEnumFeedbackTypeWithAggregatesFilterSchema: z.ZodType<Prisma.NestedEnumFeedbackTypeWithAggregatesFilter> = z.object({
  equals: z.lazy(() => FeedbackTypeSchema).optional(),
  in: z.lazy(() => FeedbackTypeSchema).array().optional(),
  notIn: z.lazy(() => FeedbackTypeSchema).array().optional(),
  not: z.union([ z.lazy(() => FeedbackTypeSchema),z.lazy(() => NestedEnumFeedbackTypeWithAggregatesFilterSchema) ]).optional(),
  _count: z.lazy(() => NestedIntFilterSchema).optional(),
  _min: z.lazy(() => NestedEnumFeedbackTypeFilterSchema).optional(),
  _max: z.lazy(() => NestedEnumFeedbackTypeFilterSchema).optional()
}).strict();

export default NestedEnumFeedbackTypeWithAggregatesFilterSchema;